### 2022-11-25 #4967 @editorialbot
[PRE REVIEW]: GNS: A generalizable Graph Neural Network-based simulator for particulate and fluid modeling
**Submitting author:** @kks32 (<a href="http://orcid.org/0000-0003-2144-5562">Krishna Kumar</a>)
**Repository:** https://github.com/geoelements/gns
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.1
