### 2022-11-25 #4967 @arfon
Output of `cloc`:

```
cloc gns/
      51 text files.
      51 unique files.
       5 files ignored.

Language                     files          blank        comment           code
SVG                              6              5              5           8853
Python                          14            261            471            870
CSS                              1              0              5            853
YAML                             3              2              5            504
Markdown                        11            120              0            277
TeX                              1              0              0            131
Bourne Shell                     8             57             91            118
HTML                             1              3              1             77
Jupyter Notebook                 1              0            146             49
JavaScript                       1              8             38             37
Dockerfile                       1              0              0             10
SUM:                            48            456            762          11779
```
