### 2022-11-26 #4967 @kks32
@arfon Thank you so much. I'd like to add the GNS code has been used as the Oracle in this conference paper https://arxiv.org/pdf/2206.12683.pdf (citation 13)
