### 2022-11-27 #4967 @gkthiruvathukal
@kks32 I can see the intellectual merit of this work. As the EIC responsible for this track (DSAIS), I share @arfon's concerns about the substantial scholarly effort criterion, which I believe has not been met. In addition, the paper you mentioned (citation 13) might raise more questions than answers as this paper also seems to be about the oracle itself. That is, as I read it, your paper is describing the oracle, so I would need to know a bit more about how this JOSS submission differs from that paper, since the conference paper may end up being one of the few places where one might cite your software.

I'll also add that your code seems more like a collection of scripts and a thin layer than general software that might be used by others. The README on your landing page suggests a significant amount of configuration via environment variables is required to make use of this with rather terse explanations.

My initial inclination is to recommend rejection based on the points raised above, but I would like to give you a chance to rebut before making this recommendation.
