### 2022-11-28 7cb4e69
Remove environment variables for file/folder names
