### 2022-11-29 #4967 @gkthiruvathukal
@kks32 Ok, I still feel like the README could be a lot better written, perhaps even including some of your explanation here and really making sure someone would be able to understand how to run this code.  The purpose of a README is to explain with the hope that someone landing on the page may be inspired to consider your work.

However, I am satisfied with your response as it clarified the potential uses/users for this work. Please keep in mind that once I assign an editor and the editor assigns reviewers, the scholarly effort criterion could arise again as our editors and reviewers check for this just about every time. Unfortunately, I had to reject a submission recently where the pre-review identified scholarly effort concerns that became more evident during the actual review. I do remain concerned about this possibility even though (as I had mentioned earlier) there is clear intellectual merit to the work. However, intellectual merit and scholarly effort are not the same thing in my world.

In any event, I will begin working on assigning an editor for this submission. Please be patient as we work to get your submission edited. Thank you again for your detailed response!
