### 2022-12-20 c640867
Explain GNS
