+++ README.md
@@ -9,6 +9,8 @@
 
 > Joseph Vantassel, Texas Advanced Computing Center, UT Austin.
 
+Graph Network-based Simulator (GNS) is a framework for developing generalizable, efficient, and accurate machine learning (ML)-based surrogate models for particulate and fluid systems using Graph Neural Networks (GNNs). GNS code is a viable surrogate for numerical methods such as Material Point Method, Smooth Particle Hydrodynamics and Computational Fluid dynamics. GNS exploits distributed data parallelism to achieve multi-GPU training.
+
 ## Run GNS
 > Training
 ```shell
