### 2022-12-20 #4967 @osorensen
:wave: @WPettersson @zadehsey @olgadoronina, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
