### 2022-12-20 #4967 @osorensen
:wave: @TranzWu, given your expertise in molecular dynamics, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
