### 2023-01-05 geoelements/gns#34 @WPettersson
Missing installation instructions for non-TACC users
Your README file includes details on how to install/use gns on TACC, but doesn't have information for people who won't be using TACC. It might be very hard to write a singular script to handle all cases, but it would be good to at least have a list of required packages (preferably with a link on where to find them) and compatible version numbers. For instance, the `build_venv.sh` file pulls in cuda 11.3 - will other versions work? It's okay to say "we don't know" if you don't know, but it's helpful to also point out that it is meant to work with 11.3.
