### 2023-01-06 geoelements/gns#34 @kks32
Thanks @WPettersson. PyTorch Geometric and CUDA have only specific compatibilites as shown here
https://pytorch-geometric.readthedocs.io/en/latest/notes/installation.html#quick-start

We could add a similar table in the README to help users with installation on GPU machines.
