### 2023-01-06 geoelements/gns#34 @WPettersson
An even easier solution would be to simply link to those instructions with a sentence like "GNS uses pytorch geometric and CUDA. These packages have specific requirements, please see <link> for details on installing pytorch-geometric."
