### 2023-01-22 b8e4f4b
update renderer
Incorporate 2d and 3d gif animation and  vtk options
