+++ gns/render_rollout.py
@@ -1,32 +1,3 @@
-# Lint as: python3
-# pylint: disable=g-bad-file-header
-# Copyright 2020 DeepMind Technologies Limited. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#    http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or  implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ============================================================================
-"""Simple matplotlib rendering of a rollout prediction against ground truth.
-
-Usage (from parent directory):
-
-`python -m gns.render_rollout --rollout_path={OUTPUT_PATH}/rollout_test_1.pkl`
-
-Where {OUTPUT_PATH} is the output path passed to `train.py` in "eval_rollout"
-mode.
-
-It may require installing Tkinter with `sudo apt-get install python3.7-tk`.
-
-"""  # pylint: disable=line-too-long
-
 import pickle
 
 from absl import app
@@ -38,13 +9,15 @@ import numpy as np
 import os
 from pyevtk.hl import pointsToVTK
 
-flags.DEFINE_string("rollout_path", None, help="Path to rollout pickle file.")
+flags.DEFINE_string("rollout_dir", None, help="Directory where rollout.pkl are located")
+flags.DEFINE_string("rollout_name", None, help="Name of rollout `.pkl` file")
 flags.DEFINE_integer("step_stride", 3, help="Stride of steps to skip.")
-flags.DEFINE_boolean("block_on_show", True, help="For test purposes.")
+flags.DEFINE_enum("output_mode", "gif", ["both", "gif", "vtk"], help="Type of render output")
 
 FLAGS = flags.FLAGS
 
 TYPE_TO_COLOR = {
+    1: "red",  # for droplet
     3: "black",  # Boundary particles.
     0: "green",  # Rigid solids.
     7: "magenta",  # Goop.
@@ -53,69 +26,193 @@ TYPE_TO_COLOR = {
 }
 
 
-def main(unused_argv):
+class Render():
 
-  if not FLAGS.rollout_path:
-    raise ValueError("A `rollout_path` must be passed.")
-  with open(FLAGS.rollout_path, "rb") as file:
-    rollout_data = pickle.load(file)
+    def __init__(self, input_dir, input_name):
+        # Texts to describe rollout cases for data and render
+        rollout_cases = [
+            ["ground_truth_rollout", "MPM"], ["predicted_rollout", "GNS"]]
+        self.rollout_cases = rollout_cases
+        self.input_dir = input_dir
+        self.input_name = input_name
+        self.output_dir = input_dir
+        self.output_name = input_name
 
-  fig, axes = plt.subplots(1, 2, figsize=(10, 5))
+        # Get trajectory
+        with open(f"{self.input_dir}{self.input_name}.pkl", "rb") as file:
+            rollout_data = pickle.load(file)
+        self.rollout_data = rollout_data
+        trajectory = {}
+        for rollout_case in rollout_cases:
+            trajectory[rollout_case[0]] = np.concatenate(
+                [rollout_data["initial_positions"], rollout_data[rollout_case[0]]], axis=0
+            )
+        self.trajectory = trajectory
 
-  plot_info = []
-  for ax_i, (label, rollout_field) in enumerate(
-      [("Reality", "ground_truth_rollout"),
[…]
