### 2023-01-27 babbe16
remove  option and remove redundancy
