### 2023-01-28 a00a6e5
Update renderer (#37)
* update renderer

Incorporate 2d and 3d gif animation and  vtk options

* update readme for new render code

* remove  option and remove redundancy
