### 2023-02-18 geoelements/gns#40 @WPettersson
Functionality testing
**Describe the feature**
The few tests included do little more than test if the imports cause Python to error. Testing really should include results are accurate.

**Describe alternatives**
Include more tests that check whether the code accurately calculates what it claims to calculate.
