### 2023-02-18 geoelements/gns#41 @WPettersson
API documentation improvements 
**Describe the feature**
GNS is described in the JOSS paper as a "framework for developing generalizable, efficient, and accurate machine learning ... models". If this is the case, then the API level documentation is lacking. Some functions do have docstrings, but many functions and variables are not documented.

**Describe alternatives**
At a minimum, add docstrings to any functions that may be useful for someone trying to use GNS as a framework to develop new machine learning models. You should probably have a list somewhere of "Here are the things you need to do to implement a new model."

**Additional context**
 It's also generally nice to host a HTML extract of such documentation online somewhere. Many use readthedocs.io for this, and you can automatically extract doc strings into the HTML using continuous deployment which makes things easier.
