### 2023-02-18 #5025 @WPettersson
Just so I can track what I'm waiting on (didn't want to add chatter around the official review checklist)

Waiting on: https://github.com/geoelements/gns/issues/34

[ ] Installation: Does installation proceed as outlined in the documentation?
[ ] Functionality: Have the functional claims of the software been confirmed?
[ ] Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)
[ ] Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

Waiting on: https://github.com/geoelements/gns/issues/41

[ ] Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

Waiting on: https://github.com/geoelements/gns/issues/40

[ ] Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

Waiting on: https://github.com/geoelements/gns/issues/35

[ ] Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
[ ] State of the field: Do the authors describe how this software compares to other commonly-used packages?
[ ] References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?
