### 2023-03-23 ea61fe2
Fix the documentation KaTeX rendering
