+++ docs/_sidebar.md
@@ -1,1 +1,3 @@
 <!-- docs/_sidebar.md -->
+* GNS Theory
+  * [Method](method.md)
