+++ docs/_sidebar.md
@@ -1,3 +1,3 @@
 <!-- docs/_sidebar.md -->
 * GNS Theory
-  * [Method](method.md)
+  * [theory](theory.md)
