+++ docs/_sidebar.md
@@ -1,3 +1,2 @@
 <!-- docs/_sidebar.md -->
-* GNS Theory
-  * [theory](theory.md)
+* [GNS Theory](theory.md)
