### 2023-03-23 236a214
Refactor image location for rollout
