+++ docs/README.md
@@ -1,8 +1,5 @@
-# Graph Networks
+# Graph Network Simulator
 
-## Multi-Layer Perceptrons (MLPs)
-![mlp](img/mlp.svg)
+Graph Network-based Simulator (GNS) is a framework for developing generalizable, efficient, and accurate machine learning (ML)-based surrogate models for particulate and fluid systems using Graph Neural Networks (GNNs). GNS code is a viable surrogate for numerical methods such as Material Point Method, Smooth Particle Hydrodynamics and Computational Fluid dynamics. GNS exploits distributed data parallelism to achieve fast multi-GPU training. The GNS code can handle complex boundary conditions and multi-material interactions.  GNS is a viable surrogate for numerical models such as Material Point Method, Smooth Particle Hydrodynamics and Computational Fluid dynamics.
 
-
-## Graph Networks (GNS)
-![gns](img/gns.svg)
+![Sand rollout](img/rollout_0.gif)
