+++ docs/_sidebar.md
@@ -1,1 +1,2 @@
 <!-- docs/_sidebar.md -->
+* [GNS Theory](theory.md)
