+++ docs/theory.md
@@ -1,0 +1,196 @@
+# Method
+
+## Graph Neural Networks (GNNs) and message passing
+
+### Graphs
+Graphs are powerful means of representing interactions between physical systems. 
+A granular material media can be represented as a graph $ G=\left(V,E\right) $ consisting of a set of vertices ($\bm{v}_i \in\ V$) representing the soil grains and edges ($\bm{e}_{i,j} \in\ E$) connecting a pair of vertices ($\bm{v}_i$ and $\bm{v}_j$) representing the interaction relationship between grains. 
+We describe how graphs work by showing a simple example involving interaction between balls in a box (Figure 1a). 
+The state of the physical system (Figure 1a and 1d) can be encoded as a graph (Figure 1b and 1c). The vertices describe the balls, and the edges describe the directional interaction between them, shown as arrows in Figure 1b and 1c. 
+The state of the ball i is represented as a vertex feature vector $\bm{v}_i$ at $i$. The feature vector includes properties such as velocities, mass, and distance to the boundary. 
+The edge feature vector $\bm{e}_{i,j}$ includes the information about the interaction between balls $i$ and $j$ such as the relative distance between the balls.
+
+Graphs offer a permutation invariant form of encoding data, where the interaction between vertices is independent of the order of vertices or their position in Euclidean space. 
+Rather,  graphs represent the interactions through the edge connection, not affected by the permutation of the vertices. 
+Therefore, graphs can efficiently represent the physical state of granular flow where numerous orderless particles interact by using vertices to represent particles and edges to their interaction. 
+
+### Graph neural networks (GNNs)
+GNNs are a state-of-the-art deep learning architecture that can operate on a graph and learn the local interactions. 
+GNNs take a graph $G=\left(\bm{V},\bm{E}\right)$ at time t as an input, compute properties and propagate information through the network, termed as message passing, and output an updated graph $G^\prime=\left(\bm{V}^\prime,\bm{E}^\prime\right)$ with an identical structure, where $\bm{V}^\prime$ and $\bm{E}^\prime$ are the set of updated vertex and edge features ($\bm{v}_i^\prime$ and $\bm{e}_{i,\ j}^\prime$).
+In the balls-in-a-box example, the GNN first takes the original graph $G=\left(\bm{V},\bm{E}\right)$  (Figure 1b) that describes the current state of the physical system ($\bm{X}^t$). 
+The GNN then updates the state of the physical system through message passing, which models the exchange of energy and momentum between the balls communicating through the edges, and returns an updated graph $G^\prime=\left(\bm{V}^\prime,\bm{E}^\prime\right)$ (Figure 1c). 
+After the GNN computation, we may decode G^\prime to extract useful information related to the future state of the physical system ($\bm{X}^{t+1}$) such as the next position or acceleration of the balls (Figure 1d).
+
+![balls-in-a-box](img/balls-in-a-box.svg)
[…]
