### 2023-03-23 7adb736
Try to trigger only if paper.md or figs change
