### 2023-04-05 #5025 @osorensen
@kks32, I notice that @WPettersson has opened several [issues in the source repo](https://github.com/geoelements/gns/issues). Could you please update us on how it's going addressing these issues?
