### 2023-05-13 cc68795
pytorch 2.0 on ls6 (#42)
* update renderer

Incorporate 2d and 3d gif animation and  vtk options

* update readme for new render code

* remove  option and remove redundancy

* add method doc

* add method doc

* oneline

* rename images

* Add KaTeX

* boldsymbol

* boldsymbol

* Fix the documentation KaTeX rendering

* GNS fix sidebar in docs

* Fix sidebar link

* Refactor image location for rollout

* pytorch2.0 on ls6

* create module.sh

---------

Co-authored-by: Krishna Kumar <krishnak@utexas.edu>
