+++ meshnet/train.py
@@ -1,0 +1,308 @@
+import sys
+import os
+import glob
+import numpy as np
+import torch
+from torch.nn.parallel import DistributedDataParallel as DDP
+import torch_geometric.transforms as T
+import re
+import pickle
+from tqdm import tqdm
+
+from absl import flags
+from absl import app
+
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+from meshnet import data_loader
+from meshnet import learned_simulator
+from meshnet.noise import get_velocity_noise
+from meshnet.utils import datas_to_graph
+from meshnet.utils import NodeType
+from meshnet.utils import optimizer_to
+
+
+flags.DEFINE_enum(
+    'mode', 'train', ['train', 'valid', 'rollout'],
+    help='Train model, validation or rollout evaluation.')
+flags.DEFINE_integer('batch_size', 2, help='The batch size.')
+flags.DEFINE_string('data_path', "/work2/08264/baagee/frontera/gns-meshnet-data/gns-data/datasets/pipe-npz/", help='The dataset directory.')
+flags.DEFINE_string('model_path', "/work2/08264/baagee/frontera/gns-meshnet-data/gns-data/models/pipe-npz/", help=('The path for saving checkpoints of the model.'))
+flags.DEFINE_string('output_path', "/work2/08264/baagee/frontera/gns-meshnet-data/gns-data/rollouts/pipe-npz/", help='The path for saving outputs (e.g. rollouts).')
+flags.DEFINE_string('model_file', "latest", help=('Model filename (.pt) to resume from. Can also use "latest" to default to newest file.'))
+flags.DEFINE_string('train_state_file', "latest", help=('Train state filename (.pt) to resume from. Can also use "latest" to default to newest file.'))
+flags.DEFINE_integer("cuda_device_number", None, help="CUDA device (zero indexed), default is None so default CUDA device will be used.")
+flags.DEFINE_string('rollout_filename', "rollout", help='Name saving the rollout')
+flags.DEFINE_integer('ntraining_steps', int(1E7), help='Number of training steps.')
+flags.DEFINE_integer('nsave_steps', int(5000), help='Number of steps at which to save the model.')
+FLAGS = flags.FLAGS
+
+
+INPUT_SEQUENCE_LENGTH = 1
+noise_std = 2e-2
+node_type_embedding_size = 9
+dt = 0.01
+lr_init = 1e-4
+lr_decay_rate = 0.1
+lr_decay_steps = 5e6
+loss_report_step = 10
+
+device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
+# an instance that transforms face-based graph to edge-based graph. Edge features are auto-computed using "Cartesian" and "Distance"
+transformer = T.Compose([T.FaceToEdge(), T.Cartesian(norm=False), T.Distance(norm=False)])
+
+
+def predict(simulator: learned_simulator.MeshSimulator,
+            device: str):
+
+    # Load simulator
+    if os.path.exists(FLAGS.model_path + FLAGS.model_file):
+        simulator.load(FLAGS.model_path + FLAGS.model_file)
+    else:
+        raise Exception(f"Model does not exist at {FLAGS.model_path + FLAGS.model_file}")
+
+    simulator.to(device)
+    simulator.eval()
+
+    # Output path
+    if not os.path.exists(FLAGS.output_path):
+        os.makedirs(FLAGS.output_path)
+
+    # Use `valid`` set for eval mode if not use `test`
+    split = 'test' if FLAGS.mode == 'rollout' else 'valid'
+
+    # Load trajectory data.
+    ds = data_loader.get_data_loader_by_trajectories(path=f"{FLAGS.data_path}{split}.npz")
[…]
