### 2023-06-19 f10c596
Describe dataset in README. Add install instructions for non-TACC users. Fixes #34
