+++ README.md
@@ -12,7 +12,7 @@
 Graph Network-based Simulator (GNS) is a framework for developing generalizable, efficient, and accurate machine learning (ML)-based surrogate models for particulate and fluid systems using Graph Neural Networks (GNNs). GNS code is a viable surrogate for numerical methods such as Material Point Method, Smooth Particle Hydrodynamics and Computational Fluid dynamics. GNS exploits distributed data parallelism to achieve fast multi-GPU training. The GNS code can handle complex boundary conditions and multi-material interactions.
 
 ## Run GNS
-> Training
+> Training GNS on data
 ```shell
 python3 -m gns.train --data_path="<input-training-data-path>" --model_path="<path-to-load-save-model-file>" --output_path="<path-to-save-output>" -ntraining_steps=100
 ```
@@ -25,7 +25,7 @@ To resume training specify `model_file` and `train_state_file`:
 python3 -m gns.train --data_path="<input-training-data-path>" --model_path="<path-to-load-save-model-file>" --output_path="<path-to-save-output>"  --model_file="model.pt" --train_state_file="train_state.pt" -ntraining_steps=100
 ```
 
-> Rollout
+> Rollout prediction
 ```shell
 python3 -m gns.train --mode="rollout" --data_path="<input-data-path>" --model_path="<path-to-load-save-model-file>" --output_path="<path-to-save-output>" --model_file="model.pt" --train_state_file="train_state.pt"
 ```
@@ -42,17 +42,28 @@ The renderer also writes `.vtu` files to visualize in ParaView.
 
 ## Datasets
 
-The data loader provided with this PyTorch implementation utilizes the more general `.npz` format. The `.npz` format includes a list of
-tuples of arbitrary length where each tuple is for a different training trajectory
-and is of the form `(position, particle_type)`. `position` is a 3-D tensor of
-shape `(n_time_steps, n_particles, n_dimensions)` and `particle_type` is
-a 1-D tensor of shape `(n_particles)`.  
+We use the numpy `.npz` format for storing positional data for GNS training.  The `.npz` format includes a list of tuples of arbitrary length where each tuple corresponds to a differenet training trajectory and is of the form `(position, particle_type)`.  The data loader provides `INPUT_SEQUENCE_LENGTH` positions, set equal to six by default, to provide the GNS with the last `INPUT_SEQUENCE_LENGTH` minus one positions as input to predict the position at the next time step.  The `position` is a 3-D tensor of shape `(n_time_steps, n_particles, n_dimensions)` and `particle_type` is a 1-D tensor of shape `(n_particles)`.  
 
 The dataset contains:
 
-* Metadata file with dataset information (sequence length, dimensionality, box bounds, default connectivity radius, statistics for normalization, ...):
+* Metadata file with dataset information `(sequence length, dimensionality, box bounds, default connectivity radius, statistics for normalization, ...)`:
 
-* npz containing data for all trajectories (particle types, positions, global context, ...):
+```
+{
+  "bounds": [[0.1, 0.9], [0.1, 0.9]], 
+  "sequence_length": 320, 
+  "default_connectivity_radius": 0.015, 
+  "dim": 2, 
+  "dt": 0.0025, 
+  "vel_mean": [5.123277536458455e-06, -0.0009965205918140803], 
+  "vel_std": [0.0021978993231675805, 0.0026653552458701774], 
+  "acc_mean": [5.237611158734309e-07, 2.3633027988858656e-07], 
+  "acc_std": [0.0002582944917306106, 0.00029554531667679154]
+}
+```
+* npz containing data for all trajectories `(particle types, positions, global context, ...)`:
+
[…]
