### 2023-06-19 dc3d207
Fix loss fn eq
