+++ references.bib
@@ -71,7 +71,7 @@ year = {2022}
 author = {Vantassel, Joseph and Kumar, Krishna},
 doi = {10.17603/ds2-0phb-dg64},
 month = {10},
-title = {{Graph Network Simulator: v1.0.1}},
+title = {{Graph Network Simulator Datasets}},
 url = {https://doi.org/10.17603/ds2-0phb-dg64},
 version = {v1},
 year = {2022}
