+++ references.bib
@@ -28,6 +28,8 @@ year = {2022}
   booktitle={International Conference on Machine Learning},
   pages={8459--8468},
   year={2020},
+  doi={10.5555/3524938},
+  url={https://dl.acm.org/doi/10.5555/3524938.3525722},
   organization={PMLR}
 }
 @article{scarselli2008graph,
