### 2023-06-19 8f4681a
GNS SOTA and training
