+++ references.bib
@@ -131,3 +131,26 @@ author={
     Xiaoqiang~Zheng},
   year={2015},
 }
+@article{soga2016,
+   author = {Soga, K. and Alonso, E. and Yerro, A. and Kumar, K. and Bandara, S.},
+   title = {Trends in large-deformation analysis of landslide mass movements with particular emphasis on the material point method},
+   journal = {Géotechnique},
+   volume = {66},
+   number = {3},
+   pages = {248-273},
+   ISSN = {0016-8505
+1751-7656},
+   DOI = {10.1680/jgeot.15.LM.005},
+   year = {2016},
+   type = {Journal Article}
+}
+@article{hu2018mlsmpmcpic,
+  title={A Moving Least Squares Material Point Method with Displacement Discontinuity and Two-Way Rigid Body Coupling},
+  author={Hu, Yuanming and Fang, Yu and Ge, Ziheng and Qu, Ziyin and Zhu, Yixin and Pradhana, Andre and Jiang, Chenfanfu},
+  journal={ACM Transactions on Graphics (TOG)},
+  volume={37},
+  number={4},
+  pages={150},
+  year={2018},
+  publisher={ACM}
+}
