### 2023-06-19 geoelements/gns#40 @kks32
Thanks @WPettersson, we do have a test in CircleCI that tests if the GNS can train and predict after 10 steps of training data. `python -m gns.train --data_path=${DATA_PATH} --model_path=${MODEL_PATH} --ntraining_steps=10 --mode='train'`. Unfortunately, due to limitations of free CI services. I'm limited to 2 CPU nodes. The GNS code is built to train on GPUs and for 2 million steps to generate meaningful prediction of trajectories. The GNS training on GPU is currently not possible on a free CI service. So, we show that the GNS works and trains for a simple problem with 10 steps.
