+++ gns/data_loader.py
@@ -2,14 +2,31 @@ import torch
 import numpy as np
 
 class SamplesDataset(torch.utils.data.Dataset):
+    """Dataset of samples of trajectories.
+    
+    Each sample is a tuple of the form (positions, particle_type).
+    positions is a numpy array of shape (sequence_length, n_particles, dimension).
+    particle_type is an integer.
 
+    Args:
+        path (str): Path to dataset.
+        input_length_sequence (int): Length of input sequence.
+
+    Attributes:
+        _data (list): List of tuples of the form (positions, particle_type).
+        _dimension (int): Dimension of the data.
+        _input_length_sequence (int): Length of input sequence.
+        _data_lengths (list): List of lengths of trajectories in the dataset.
+        _length (int): Total number of samples in the dataset.
+        _precompute_cumlengths (np.array): Precomputed cumulative lengths of trajectories in the dataset.
+    """
     def __init__(self, path, input_length_sequence):
         super().__init__()
         # load dataset stored in npz format
         # data is loaded as dict of tuples
         # of the form (positions, particle_type)
         # convert to list of tuples
-        # TODO (jpv): allow_pickle=True is potential security risk. See docs.
+        # TODO: allow_pickle=True is potential security risk. See docs.
         self._data = [item for _, item in np.load(path, allow_pickle=True).items()]
         
         # length of each trajectory in the dataset
@@ -26,9 +43,22 @@ class SamplesDataset(torch.utils.data.Dataset):
         self._precompute_cumlengths = np.array(self._precompute_cumlengths, dtype=int)
 
     def __len__(self):
+        """Return length of dataset.
+        
+        Returns:
+            int: Length of dataset.
+        """
         return self._length
 
     def __getitem__(self, idx):
+        """Returns a training example from the dataset.
+        
+        Args:
+            idx (int): Index of training example.
+
+        Returns:
+            tuple: Tuple of the form ((positions, particle_type, n_particles_per_example), label).
+        """
         # Select the trajectory immediately before
         # the one that exceeds the idx
         # (i.e., the one in which idx resides).
@@ -48,6 +78,14 @@ class SamplesDataset(torch.utils.data.Dataset):
         return ((positions, particle_type, n_particles_per_example), label)
 
 def collate_fn(data):
+    """Collate function for SamplesDataset.
+
+    Args:
+        data (list): List of tuples of the form ((positions, particle_type, n_particles_per_example), label).
+
+    Returns:
+        tuple: Tuple of the form ((positions, particle_type, n_particles_per_example), label).    
+    """
     position_list = []
     particle_type_list = []
     n_particles_per_example_list = []
@@ -68,7 +106,11 @@ def collate_fn(data):
         )
 
 class TrajectoriesDataset(torch.utils.data.Dataset):
+    """Dataset of trajectories.
 
+    Each trajectory is a tuple of the form (positions, particle_type).
+    positions is a numpy array of shape (sequence_length, n_particles, dimension).
+    """
     def __init__(self, path):
         super().__init__()
         # load dataset stored in npz format
@@ -81,9 +123,22 @@ class TrajectoriesDataset(torch.utils.data.Dataset):
         self._length = len(self._data)
 
     def __len__(self):
+        """Return length of dataset.
+        
+        Returns:
+            int: Length of dataset.
+        """
         return self._length
 
     def __getitem__(self, idx):
+        """Returns a training example from the dataset.
+        
+        Args:
+            idx (int): Index of training example.
+            
+        Returns:
+            tuple: Tuple of the form (positions, particle_type).
+        """
         positions, _particle_type = self._data[idx]
[…]
