### 2023-06-19 11419e7
WIP: Test graph network
