+++ .circleci/config.yml
@@ -20,6 +20,7 @@ jobs:
             conda install -c conda-forge matplotlib-base -y
             conda install -c conda-forge pyevtk -y
             python3 test/test_data_loader.py
+            python3 test/test_graph_network.py
             python -m gns.train --data_path=${DATA_PATH} --model_path=${MODEL_PATH} --ntraining_steps=10 --mode='train'
 
 workflows:
