### 2023-06-19 16d7dc4
Fix test case graph network
