### 2023-06-19 aa44f25
Fix shape issue in test
