### 2023-06-19 d39ca4d
Test graph network
