### 2023-06-19 1a92849
Merge graph network test (#44)
* WIP: Test graph network

* Fix filename graph network test

* Fix test case graph network

* Fix shape issue in test

* Test graph network
