+++ test/test_graph_newtork.py
@@ -1,6 +1,0 @@
-import unittest
-
-class TestGraphNetwork(unittest.TestCase):
-
-    def test_init(self):
-        self.assertTrue(True)
