+++ .circleci/config.yml
@@ -19,8 +19,10 @@ jobs:
             conda install -c conda-forge dm-tree -y
             conda install -c conda-forge matplotlib-base -y
             conda install -c conda-forge pyevtk -y
+            conda install -c conda-forge pytest -y
             python3 test/test_data_loader.py
             python3 test/test_graph_network.py
+            pytest test/test_learned_simulator.py
             python -m gns.train --data_path=${DATA_PATH} --model_path=${MODEL_PATH} --ntraining_steps=10 --mode='train'
 
 workflows:
