+++ test/test_learned_simulator.py
@@ -10,60 +10,68 @@ sys.path.append(parent_dir)
 
 from gns.learned_simulator import LearnedSimulator
 
-
-def test_learned_simulator():
-    # Define model parameters
+@pytest.fixture
+def model():
     particle_dimensions = 2
     nnode_in = 30
     nedge_in = 3
     latent_dim = 128
-    nmessage_passing_steps = 5
+    nmessage_passing_steps = 2
     nmlp_layers = 2
-    mlp_hidden_dim = 64
-    connectivity_radius = 1.0
-    boundaries = torch.tensor([[0.0, 5.0], [0.0, 5.0]])
+    mlp_hidden_dim = 128
+    connectivity_radius = 0.1
+    boundaries = np.array([[0.0, 1.0], [0.0, 1.0]])
     normalization_stats = {
-        "acceleration": {"mean": torch.tensor([0.0, 0.0]), "std": torch.tensor([1.0, 1.0])},
-        "velocity": {"mean": torch.tensor([0.0, 0.0]), "std": torch.tensor([1.0, 1.0])}
+        "acceleration": {"mean": 0.0, "std": 1.0},
+        "velocity": {"mean": 0.0, "std": 1.0}
     }
     nparticle_types = 1
     particle_type_embedding_size = 16
     device = "cpu"
 
-    # Instantiate model
-    model = LearnedSimulator(
-        particle_dimensions,
-        nnode_in,
-        nedge_in,
-        latent_dim,
-        nmessage_passing_steps,
-        nmlp_layers,
-        mlp_hidden_dim,
-        connectivity_radius,
-        boundaries.numpy(),
-        normalization_stats,
-        nparticle_types,
-        particle_type_embedding_size,
-        device
+    return LearnedSimulator(
+        particle_dimensions=particle_dimensions,
+        nnode_in=nnode_in,
+        nedge_in=nedge_in,
+        latent_dim=latent_dim,
+        nmessage_passing_steps=nmessage_passing_steps,
+        nmlp_layers=nmlp_layers,
+        mlp_hidden_dim=mlp_hidden_dim,
+        connectivity_radius=connectivity_radius,
+        boundaries=boundaries,
+        normalization_stats=normalization_stats,
+        nparticle_types=nparticle_types,
+        particle_type_embedding_size=particle_type_embedding_size,
+        device=device
     )
 
-    # Create a sequence of positions
-    position_sequence = torch.tensor([[[1.0, 2.0], [1.1, 2.1], [1.2, 2.2], [1.3, 2.3], [1.4, 2.4], [1.5, 2.5]]])
-    nparticles_per_example = torch.tensor([1])
-    particle_types = torch.tensor([0])
 
-    # Check if predict_positions works
+@pytest.fixture
+def particles():
+    current_positions = torch.tensor([[0.5, 0.5], [0.6, 0.6]])
+    position_sequence = torch.tensor(
+        [
+            [[0.3, 0.3], [0.4, 0.4], [0.5, 0.5]],
+            [[0.4, 0.4], [0.5, 0.5], [0.6, 0.6]]
+        ]
+    )
+    nparticles_per_example = torch.tensor([2])
+    particle_types = torch.tensor([0, 0])
+
+    return current_positions, position_sequence, nparticles_per_example, particle_types
+
+
+def test_predict_positions(model, particles):
+    current_positions, position_sequence, nparticles_per_example, particle_types = particles
+
     next_positions = model.predict_positions(
-        position_sequence, nparticles_per_example, particle_types)
-    assert next_positions.shape == (1, 2)
+        current_positions, nparticles_per_example, particle_types)
 
-    # Check if predict_accelerations works
-    predicted_normalized_acceleration, target_normalized_acceleration = model.predict_accelerations(
[…]
