+++ test/test_learned_simulator.py
@@ -46,7 +46,7 @@ def test_encoder_preprocessor(simulator):
     node_features, edge_index, edge_features = simulator._encoder_preprocessor(
         position_sequence, nparticles_per_example, particle_types)
 
-    assert node_features.shape == (2, 30)
+    assert node_features.shape == (2, 14)
     assert edge_index.shape == (2, 2)  # one edge between the 2 particles
     assert edge_features.shape == (2, 3)  # one edge with 3 features
 
