### 2023-06-19 b4419fd
Test encoder processor in learned simulator
