### 2023-06-19 7174215
Test noise utils
