### 2023-06-19 4350853
Print arrays
