+++ test/test_noise_utils.py
@@ -10,30 +10,34 @@ sys.path.append(parent_dir)
 from gns.noise_utils import get_random_walk_noise_for_position_sequence
 from gns.learned_simulator import time_diff
 
-@pytest.mark.parametrize("nparticles, dim, noise_std_last_step", [(10, 2, 0.1), (20, 3, 0.5)])
-def test_get_random_walk_noise_for_position_sequence(nparticles, dim, noise_std_last_step):
-    position_sequence = torch.randn(nparticles, 6, dim)
-    position_sequence_noise = get_random_walk_noise_for_position_sequence(
-        position_sequence, noise_std_last_step)
+def test_get_random_walk_noise_for_position_sequence():
+    # Define input tensor for position_sequence
+    position_sequence = torch.rand((10, 6, 3))
 
-    assert position_sequence_noise.shape == position_sequence.shape, \
-        "Output tensor has incorrect shape."
+    # Define standard deviation for last step
+    noise_std_last_step = 0.2
 
-    velocity_sequence_noise = time_diff(position_sequence_noise)
-    # Compute the standard deviation of the noise in the last velocity
-    computed_noise_std_last_step = torch.std(velocity_sequence_noise[:, -1, :])
-    print("computed_noise_std_last_step: ", computed_noise_std_last_step)
-    print("noise_std_last_step: ", noise_std_last_step)
+    # Call the function with the defined inputs
+    output = get_random_walk_noise_for_position_sequence(position_sequence, noise_std_last_step)
 
-    """
-    np.testing.assert_allclose(
-        computed_noise_std_last_step.numpy(), 
-        noise_std_last_step, 
-        atol=1e-3, 
-        err_msg="Standard deviation of the noise in the last step does not match the input."
-    )
-    """
+    # Assert that the function doesn't fail and returns a tensor of the same shape
+    assert output.shape == position_sequence.shape
+
+    # Assert that the standard deviation of the last step is close to the desired value
+    # We use np.isclose to deal with floating point imprecision
+    assert np.isclose(output[:, -1].std().item(), noise_std_last_step, atol=0.05)
 
     # Check that the first position has no noise
-    assert torch.all(position_sequence_noise[:, 0, :] == 0), \
+    assert torch.all(output[:, 0, :] == 0), \
         "The first position is expected to have no noise."
+
+@pytest.mark.parametrize("shape", [
+    (10, 6, 3),
+    (1, 5, 3),
+    (100, 7, 2)
+])
+def test_shapes(shape):
+    position_sequence = torch.rand(shape)
+    noise_std_last_step = 0.1
+    output = get_random_walk_noise_for_position_sequence(position_sequence, noise_std_last_step)
+    assert output.shape == position_sequence.shape
