### 2023-06-19 9de1515
Run pytest for graph network
