### 2023-06-25 299d90c
Fix rollout related bugs (see issue #32)
