### 2023-06-26 78a91d9
Refactor meshnet simulation animation
