+++ docs/noise_utils.rst
@@ -1,0 +1,7 @@
+noise\_utils module
+===================
+
+.. automodule:: noise_utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
