+++ docs/learned_simulator.rst
@@ -1,0 +1,7 @@
+learned\_simulator module
+=========================
+
+.. automodule:: learned_simulator
+   :members:
+   :undoc-members:
+   :show-inheritance:
