+++ docs/reading_utils.rst
@@ -1,0 +1,7 @@
+reading\_utils module
+=====================
+
+.. automodule:: reading_utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
