+++ docs/metadata.rst
@@ -1,0 +1,7 @@
+metadata module
+===============
+
+.. automodule:: metadata
+   :members:
+   :undoc-members:
+   :show-inheritance:
