+++ docs/index.rst
@@ -1,0 +1,20 @@
+.. GNS documentation master file, created by
+   sphinx-quickstart on Tue Jun 27 06:36:56 2023.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to GNS's documentation!
+===============================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
