+++ docs/data_loader.rst
@@ -1,0 +1,7 @@
+data\_loader module
+===================
+
+.. automodule:: data_loader
+   :members:
+   :undoc-members:
+   :show-inheritance:
