+++ docs/distribute.rst
@@ -1,0 +1,7 @@
+distribute module
+=================
+
+.. automodule:: distribute
+   :members:
+   :undoc-members:
+   :show-inheritance:
