+++ docs/modules.rst
@@ -1,0 +1,15 @@
+gns
+===
+
+.. toctree::
+   :maxdepth: 4
+
+   data_loader
+   distribute
+   graph_network
+   learned_simulator
+   metadata
+   noise_utils
+   reading_utils
+   render_rollout
+   train
