+++ docs/graph_network.rst
@@ -1,0 +1,7 @@
+graph\_network module
+=====================
+
+.. automodule:: graph_network
+   :members:
+   :undoc-members:
+   :show-inheritance:
