### 2023-06-27 cf3da62
Fix numpy version in requirements.txt
