### 2023-07-07 #5025 @osorensen
Thanks a lot for your work @WPettersson! I think the word count for the paper is fine, particularly since the number of references contributes to inflating it a bit.
