### 2023-07-07 #5025 @osorensen
👋 @TranzWu, could you please take a new look at [your checklist](https://github.com/openjournals/joss-reviews/issues/5025#issuecomment-1438806321) and see if further of the boxes can be checked? If you have any particular points about the submission, please mention them here or open an issue in the [source repository](https://github.com/geoelements/gns).
