### 2023-07-27 #5025 @archermarx
> 👋 @archermarx @ashwinvis, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Apologies for the delay on this. I was at a conference. I am happy to review this submission!
