### 2023-07-27 #5025 @archermarx
I've gone through the paper and have no major edits. Tomorrow or monday, I will install and test the software to verify it works as intended.
