### 2023-08-03 #5025 @osorensen
@archermarx, thanks a lot for going through the paper so quickly. Have you had a chance to look at the software yet? It's absolutely fine if you haven't, but if you are able to do so within the next two weeks, it would be very much appreciated.
