### 2023-08-03 #5025 @archermarx
Yes! I'll try to take a look today

On Thu, Aug 3, 2023, 9:01 AM Øystein Sørensen ***@***.***>
wrote:

> @archermarx <https://github.com/archermarx>, thanks a lot for going
> through the paper so quickly. Have you had a chance to look at the software
> yet? It's absolutely fine if you haven't, but if you are able to do so
> within the next two weeks, it would be very much appreciated.
>
