### 2023-08-07 #5025 @archermarx
I am encountering an error installing the code. After following the instructions to install pytorch_geometric

```
pip install pytorch_geometric
```

I ran

```
pip install -R requirements.txt
```
in the `gns directory`, and after some time I got the following error:

```
PS C:\Users\thoma\gns> pip install -r requirements.txt
Collecting absl-py (from -r requirements.txt (line 1))
  Using cached absl_py-1.4.0-py3-none-any.whl (126 kB)
Collecting autopep8 (from -r requirements.txt (line 2))
  Using cached autopep8-2.0.2-py2.py3-none-any.whl (45 kB)
Collecting numpy==1.23.1 (from -r requirements.txt (line 3))
  Using cached numpy-1.23.1-cp310-cp310-win_amd64.whl (14.6 MB)
Collecting dm-tree (from -r requirements.txt (line 4))
  Using cached dm_tree-0.1.8-cp310-cp310-win_amd64.whl (101 kB)
Requirement already satisfied: matplotlib in c:\users\thoma\appdata\local\packages\pythonsoftwarefoundation.python.3.10_qbz5n2kfra8p0\localcache\local-packages\python310\site-packages (from -r requirements.txt (line 5)) (3.6.1)
Collecting pyevtk (from -r requirements.txt (line 6))
  Using cached pyevtk-1.6.0-py3-none-any.whl (20 kB)
Requirement already satisfied: pytest in c:\users\thoma\appdata\local\packages\pythonsoftwarefoundation.python.3.10_qbz5n2kfra8p0\localcache\local-packages\python310\site-packages (from -r requirements.txt (line 7)) (7.1.3)
Requirement already satisfied: torch in c:\users\thoma\appdata\local\packages\pythonsoftwarefoundation.python.3.10_qbz5n2kfra8p0\localcache\local-packages\python310\site-packages (from -r requirements.txt (line 8)) (2.0.1)
Requirement already satisfied: torch_geometric in c:\users\thoma\appdata\local\packages\pythonsoftwarefoundation.python.3.10_qbz5n2kfra8p0\localcache\local-packages\python310\site-packages (from -r requirements.txt (line 9)) (2.3.1)
Collecting torch_sparse (from -r requirements.txt (line 10))
  Using cached torch_sparse-0.6.17.tar.gz (209 kB)
  Installing build dependencies ... done
  Getting requirements to build wheel ... error
  error: subprocess-exited-with-error

  × Getting requirements to build wheel did not run successfully.
  │ exit code: 1
  ╰─> [17 lines of output]
      Traceback (most recent call last):
        File "C:\Users\thoma\AppData\Local\Packages\PythonSoftwareFoundation.Python.3.10_qbz5n2kfra8p0\LocalCache\local-packages\Python310\site-packages\pip\_vendor\pyproject_hooks\_in_process\_in_process.py", line 353, in <module>
          main()
        File "C:\Users\thoma\AppData\Local\Packages\PythonSoftwareFoundation.Python.3.10_qbz5n2kfra8p0\LocalCache\local-packages\Python310\site-packages\pip\_vendor\pyproject_hooks\_in_process\_in_process.py", line 335, in main
          json_out['return_val'] = hook(**hook_input['kwargs'])
        File "C:\Users\thoma\AppData\Local\Packages\PythonSoftwareFoundation.Python.3.10_qbz5n2kfra8p0\LocalCache\local-packages\Python310\site-packages\pip\_vendor\pyproject_hooks\_in_process\_in_process.py", line 118, in get_requires_for_build_wheel
          return hook(config_settings)
        File "C:\Users\thoma\AppData\Local\Temp\pip-build-env-2f8616_v\overlay\Lib\site-packages\setuptools\build_meta.py", line 341, in get_requires_for_build_wheel
          return self._get_build_requires(config_settings, requirements=['wheel'])
        File "C:\Users\thoma\AppData\Local\Temp\pip-build-env-2f8616_v\overlay\Lib\site-packages\setuptools\build_meta.py", line 323, in _get_build_requires
          self.run_setup()
        File "C:\Users\thoma\AppData\Local\Temp\pip-build-env-2f8616_v\overlay\Lib\site-packages\setuptools\build_meta.py", line 487, in run_setup
          super(_BuildMetaLegacyBackend,
        File "C:\Users\thoma\AppData\Local\Temp\pip-build-env-2f8616_v\overlay\Lib\site-packages\setuptools\build_meta.py", line 338, in run_setup
          exec(code, locals())
        File "<string>", line 8, in <module>
      **ModuleNotFoundError: No module named 'torch'**
      [end of output]

error: subprocess-exited-with-error

× Getting requirements to build wheel did not run successfully.
│ exit code: 1
╰─> See above for output.

```

As can be seen, I have both pytorch and pytorch_geometric installed, so I'm not sure what's causing this.
