### 2023-08-07 #5025 @kks32
@archermarx Sorry about that. If you are on Linux, could you try the following for installation. PyTorch Geometric always creates some issues:

```
conda install pytorch torchvision torchaudio cpuonly -c pytorch
conda install pyg -c pyg
conda install -c conda-forge numpy dm-tree matplotlib-base pyevtk
```
Please feel free to open an issue on [GNS](https://github.com/geoelements/gns/issues/new) if you are having issues so we can resolve them. Thank you for reviewing the code!

I have updated instructions on the repo too.
