### 2023-08-08 #5025 @archermarx
So i followed the new conda installation instructions and didn't get any errors. I then attempted to follow your instructions for training. For context, I have the gns repository and the gns-sample repository in the same directory. From the gns repository, I run:

```
python3 -m gns.train --data_path="../gns-sample/WaterDropSample/dataset/test.npz" --output_path="output" --model_path="models" -ntraining_steps=1000
```

The terminal outputs `world_size=0` and then exits. There does not seem to be any output produced. There is no `output` file. and the `models` directory is empty.

Also, how do I run the automated tests?
