### 2023-08-09 b80d8a2
Dev/enable cpu (#49)
* enable cpu training

* Update read me

* Remove grad checkpointing from this PR

---------

Co-authored-by: Krishna Kumar <3963513+kks32@users.noreply.github.com>
