+++ gns/train.py
@@ -146,7 +146,7 @@ def predict(device: str, FLAGS, flags, world_size):
       example_rollout['metadata'] = metadata
       print("Predicting example {} loss: {}".format(example_i, loss.mean()))
       eval_loss.append(torch.flatten(loss))
-      
+
       # Save rollout in testing
       if FLAGS.mode == 'rollout':
         example_rollout['metadata'] = metadata
@@ -179,13 +179,20 @@ def train(rank, flags, world_size):
     rank: local rank
     world_size: total number of ranks
   """
-  distribute.setup(rank, world_size)
+  if type(rank) == int:
+    distribute.setup(rank, world_size)
+  else:
+    rank = torch.device("cpu")
 
   metadata = reading_utils.read_metadata(flags["data_path"])
-  serial_simulator = _get_simulator(metadata, flags["noise_std"], flags["noise_std"], rank)
 
-  simulator = DDP(serial_simulator.to(rank), device_ids=[rank], output_device=rank)
-  optimizer = torch.optim.Adam(simulator.parameters(), lr=flags["lr_init"]*world_size)
+  if type(rank) == int:
+    serial_simulator = _get_simulator(metadata, flags["noise_std"], flags["noise_std"], rank)
+    simulator = DDP(serial_simulator.to(rank), device_ids=[rank], output_device=rank)
+    optimizer = torch.optim.Adam(simulator.parameters(), lr=flags["lr_init"]*world_size)
+  else:
+    simulator = _get_simulator(metadata, flags["noise_std"], flags["noise_std"], rank)
+    optimizer = torch.optim.Adam(simulator.parameters(), lr=flags["lr_init"] * world_size)
   step = 0
 
   # If model_path does exist and model_file and train_state_file exist continue training.
@@ -216,24 +223,33 @@ def train(rank, flags, world_size):
       optimizer_to(optimizer, rank)
       # set global train state
       step = train_state["global_train_state"].pop("step")
- 
+
     else:
       msg = f'Specified model_file {flags["model_path"] + flags["model_file"]} and train_state_file {flags["model_path"] + flags["train_state_file"]} not found.'
-      raise FileNotFoundError(msg) 
+      raise FileNotFoundError(msg)
 
   simulator.train()
   simulator.to(rank)
 
-  dl = distribute.get_data_distributed_dataloader_by_samples(path=f'{flags["data_path"]}train.npz',
-                                                             input_length_sequence=INPUT_SEQUENCE_LENGTH,
-                                                             batch_size=flags["batch_size"],
-                                                            )
+  if type(rank) == int:
+    dl = distribute.get_data_distributed_dataloader_by_samples(path=f'{flags["data_path"]}train.npz',
+                                                               input_length_sequence=INPUT_SEQUENCE_LENGTH,
+                                                               batch_size=flags["batch_size"],
+                                                               )
+  else:
+    dl = data_loader.get_data_loader_by_samples(path=f'{flags["data_path"]}train.npz',
+                                                input_length_sequence=INPUT_SEQUENCE_LENGTH,
+                                                batch_size=flags["batch_size"],
+                                                )
 
   print(f"rank = {rank}, cuda = {torch.cuda.is_available()}")
   not_reached_nsteps = True
   try:
     while not_reached_nsteps:
-      torch.distributed.barrier()
+      if type(rank) == int:
+        torch.distributed.barrier()
+      else:
+        pass
       for ((position, particle_type, n_particles_per_example), labels) in dl:
         position.to(rank)
[…]
