### 2023-08-09 9ec455d
Print test paths and sample folder
