### 2023-08-09 e5d9914
Fix path to dataset in CI test
