### 2023-08-09 a97acbd
Fix CPU CI testing (#50)
* Print test paths and sample folder

* Fix path to dataset in CI test

* Refactor training loop for CPU and GPU and cleanup only for GPU
