+++ .circleci/config.yml
@@ -29,6 +29,7 @@ jobs:
             python -m gns.train --data_path=${DATA_PATH} --model_path=${MODEL_PATH} --ntraining_steps=10
             echo "Predict rollout"
             python3 -m gns.train --mode="rollout" --data_path=${DATA_PATH} --model_path=${MODEL_PATH} --output_path=${ROLLOUT_PATH} --model_file="model.pt" --train_state_file="train_state.pt"
+            
 
 workflows:
     version: 2
