### 2023-08-09 839ba27
Try removing .git at clone step
