+++ .circleci/config.yml
@@ -11,7 +11,7 @@ jobs:
           command: |
             TMP_DIR="./gns-sample"
             DATASET_NAME="WaterDropSample"
-            git clone https://github.com/geoelements/gns-sample.git 
+            git clone https://github.com/geoelements/gns-sample
             mkdir -p ${TMP_DIR}/${DATASET_NAME}/models/
             mkdir -p ${TMP_DIR}/${DATASET_NAME}/rollout/
             DATA_PATH="${TMP_DIR}/${DATASET_NAME}/dataset/"
@@ -29,7 +29,7 @@ jobs:
             python -m gns.train --data_path=${DATA_PATH} --model_path=${MODEL_PATH} --ntraining_steps=10
             echo "Predict rollout"
             python3 -m gns.train --mode="rollout" --data_path=${DATA_PATH} --model_path=${MODEL_PATH} --output_path=${ROLLOUT_PATH} --model_file="model.pt" --train_state_file="train_state.pt"
-            
+
 
 workflows:
     version: 2
