### 2023-08-09 65f0947
Debug to see if model files are written
