### 2023-08-09 #5025 @kks32
Thanks @archermarx for testing it. Apologies, for the issue. This is because we were set-up to run primarily on distributed GPU systems. We have updated the code so you can run it on cpus without running into issues. Assuming you will clone gns-sample to test run in the gns folder itself. Here are the steps:

```
git clone https://github.com/geoelements/gns-sample

TMP_DIR="./gns-sample"
DATASET_NAME="WaterDropSample"

mkdir -p ${TMP_DIR}/${DATASET_NAME}/models/
mkdir -p ${TMP_DIR}/${DATASET_NAME}/rollout/

DATA_PATH="${TMP_DIR}/${DATASET_NAME}/dataset/"
MODEL_PATH="${TMP_DIR}/${DATASET_NAME}/models/"
ROLLOUT_PATH="${TMP_DIR}/${DATASET_NAME}/rollout/"

python -m gns.train --data_path=${DATA_PATH} --model_path=${MODEL_PATH} --ntraining_steps=10
```

This should train the gns code for 10 steps. You can see this particular CI testing done in [circleci](https://github.com/geoelements/gns/blob/main/.circleci/config.yml). Since we use random seeding, the loss maybe different at each time.

We also have unit tests which you can run by doing: `pytest test/` in the `gns` folder. This is also done in CircleCI.

I appreciate your time and feedback! Thank you. Apologies for issues.
