### 2023-08-09 405335f
Debug model path for rollout
