### 2023-08-09 f74a005
Save steps embedded in cpu mode as well
