### 2023-08-09 aa3a07d
Save model file in CPU
