+++ gns/train.py
@@ -300,7 +300,10 @@ def train(rank, flags, world_size):
           print(f'Training step: {step}/{flags["ntraining_steps"]}. Loss: {loss}.')
           # Save model state
           if step % flags["nsave_steps"] == 0:
-            simulator.module.save(flags["model_path"] + 'model-'+str(step)+'.pt')
+            if torch.device("cpu"):
+              simulator.save(flags["model_path"] + 'model-'+str(step)+'.pt')
+            else:
+              simulator.module.save(flags["model_path"] + 'model-'+str(step)+'.pt')
             train_state = dict(optimizer_state=optimizer.state_dict(), global_train_state={"step":step})
             torch.save(train_state, f'{flags["model_path"]}train_state-{step}.pt')
 
@@ -314,8 +317,11 @@ def train(rank, flags, world_size):
   except KeyboardInterrupt:
     pass
 
-  if rank == 0:
-    simulator.module.save(flags["model_path"] + 'model-'+str(step)+'.pt')
+  if rank == 0 or torch.device("cpu"):
+    if torch.device("cpu"):
+      simulator.save(flags["model_path"] + 'model-'+str(step)+'.pt')
+    else:
+      simulator.module.save(flags["model_path"] + 'model-'+str(step)+'.pt')
     train_state = dict(optimizer_state=optimizer.state_dict(), global_train_state={"step":step})
     torch.save(train_state, f'{flags["model_path"]}train_state-{step}.pt')
 
