### 2023-08-09 1965939
Add save files for cpu version (#51)
* Print test paths and sample folder

* Fix path to dataset in CI test

* Refactor training loop for CPU and GPU and cleanup only for GPU

* Test rollout prediction

* Try for EOF issue

* Try removing .git at clone step

* Fix rollout path

* Debug to see if model files are written

* Debug model path for rollout

* Save steps embedded in cpu mode as well

* Save model file in CPU

* Add instructions to test
