+++ .circleci/config.yml
@@ -11,10 +11,12 @@ jobs:
           command: |
             TMP_DIR="./gns-sample"
             DATASET_NAME="WaterDropSample"
-            git clone https://github.com/geoelements/gns-sample.git 
+            git clone https://github.com/geoelements/gns-sample
+            mkdir -p ${TMP_DIR}/${DATASET_NAME}/models/
+            mkdir -p ${TMP_DIR}/${DATASET_NAME}/rollout/
             DATA_PATH="${TMP_DIR}/${DATASET_NAME}/dataset/"
             MODEL_PATH="${TMP_DIR}/${DATASET_NAME}/models/"
-            mkdir -p ${TMP_DIR}/${DATASET_NAME}/models/
+            ROLLOUT_PATH="${TMP_DIR}/${DATASET_NAME}/rollout/"
             conda install -c anaconda absl-py -y
             conda install -c conda-forge numpy -y
             conda install -c conda-forge dm-tree -y
@@ -25,7 +27,9 @@ jobs:
             echo "Test paths: ${DATA_PATH} ${MODEL_PATH}"
             ls
             python -m gns.train --data_path=${DATA_PATH} --model_path=${MODEL_PATH} --ntraining_steps=10
-
+            echo "Predict rollout"
+            ls ./gns-sample/WaterDropSample/models/
+            
 workflows:
     version: 2
     build:
