### 2023-08-09 #5025 @archermarx
@kks32 thanks for your responsiveness on this. I think we're almost there. I am now encountering this error:

```
rank = cpu, cuda = False
Traceback (most recent call last):
  File "/home/marksta/miniconda3/lib/python3.10/runpy.py", line 196, in _run_module_as_main
    return _run_code(code, main_globals, None,
  File "/home/marksta/miniconda3/lib/python3.10/runpy.py", line 86, in _run_code
    exec(code, run_globals)
  File "/home/marksta/gns/gns/train.py", line 425, in <module>
    app.run(main)
  File "/home/marksta/miniconda3/lib/python3.10/site-packages/absl/app.py", line 308, in run
    _run_main(main, args)
  File "/home/marksta/miniconda3/lib/python3.10/site-packages/absl/app.py", line 254, in _run_main
    sys.exit(main(argv))
  File "/home/marksta/gns/gns/train.py", line 414, in main
    train(device, myflags, world_size)
  File "/home/marksta/gns/gns/train.py", line 274, in train
    pred_acc, target_acc = simulator.predict_accelerations(
  File "/home/marksta/gns/gns/learned_simulator.py", line 280, in predict_accelerations
    node_features, edge_index, edge_features = self._encoder_preprocessor(
  File "/home/marksta/gns/gns/learned_simulator.py", line 127, in _encoder_preprocessor
    senders, receivers = self._compute_graph_connectivity(
  File "/home/marksta/gns/gns/learned_simulator.py", line 96, in _compute_graph_connectivity
    edge_index = radius_graph(
  File "/home/marksta/miniconda3/lib/python3.10/site-packages/torch_geometric/nn/pool/__init__.py", line 210, in radius_graph
    return torch_cluster.radius_graph(x, r, batch, loop, max_num_neighbors,
AttributeError: 'NoneType' object has no attribute 'radius_graph'
```

I get the same error when running tests. I've attached the full test log to this post.

[testlog.txt](https://github.com/openjournals/joss-reviews/files/12304304/testlog.txt)
