### 2023-08-09 #5025 @kks32
Hi @archermarx Thank you for your patience. The issue is related to Pytorch geometric package and its different packagings. Could you please install torch cluster using `conda install pytorch-cluster -c pyg` and then try the steps here: https://github.com/openjournals/joss-reviews/issues/5025#issuecomment-1671457463

Appreciate your patience.
