### 2023-08-14 #5025 @archermarx
Ok, I was able to successfully train for 1000 steps, and the loss decreased as expected. One final thing. Running tests still gives me an error, which looks like it has something to do with numpy. Here's the error text:

```
test/test_data_loader.py EE                                                 [ 33%]
test/test_graph_network.py .                                                [ 50%]
test/test_learned_simulator.py .                                            [ 66%]
test/test_noise_utils.py ..                                                 [100%]

===================================== ERRORS ======================================
_____________________ ERROR at setup of test_samples_dataset ______________________

temp_dir = '/tmp/tmpkm93oz5k'

    @pytest.fixture
    def dummy_data(temp_dir):
        # Create a dummy dataset
        dummy_data = [(np.random.rand(10, 3, 2), i % 3) for i in range(5)]
        data_path = os.path.join(temp_dir, 'data.npz')
>       np.savez(data_path, *dummy_data)

test/test_data_loader.py:25:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
../miniconda3/lib/python3.10/site-packages/numpy/lib/npyio.py:639: in savez
    _savez(file, args, kwds, False)
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

file = '/tmp/tmpkm93oz5k/data.npz'
args = ((array([[[2.83082370e-01, 2.42428869e-01],
        [8.70220391e-01, 8.13641733e-01],
        [1.34136507e-01, 5.43269...803545]],

       [[0.9102955 , 0.96535921],
        [0.87278961, 0.91629958],
        [0.72771614, 0.51367583]]]), 1))
kwds = {'arr_0': (array([[[2.83082370e-01, 2.42428869e-01],
        [8.70220391e-01, 8.13641733e-01],
        [1.34136507e-01...3]],

       [[0.72921478, 0.46569833],
        [0.89067798, 0.91844627],
        [0.43296379, 0.66740189]]]), 0), ...}
compress = False, allow_pickle = True, pickle_kwargs = None

    def _savez(file, args, kwds, compress, allow_pickle=True, pickle_kwargs=None):
        # Import is postponed to here since zipfile depends on gzip, an optional
        # component of the so-called standard library.
        import zipfile

        if not hasattr(file, 'write'):
            file = os_fspath(file)
            if not file.endswith('.npz'):
                file = file + '.npz'

        namedict = kwds
        for i, val in enumerate(args):
            key = 'arr_%d' % i
            if key in namedict.keys():
                raise ValueError(
                    "Cannot use un-named variables and keyword %s" % key)
            namedict[key] = val

        if compress:
            compression = zipfile.ZIP_DEFLATED
        else:
            compression = zipfile.ZIP_STORED

        zipf = zipfile_factory(file, mode="w", compression=compression)

        for key, val in namedict.items():
            fname = key + '.npy'
>           val = np.asanyarray(val)
E           ValueError: setting an array element with a sequence. The requested array has an inhomogeneous shape after 1 dimensions. The detected shape was (2,) + inhomogeneous part.

../miniconda3/lib/python3.10/site-packages/numpy/lib/npyio.py:740: ValueError
___________________ ERROR at setup of test_trajectories_dataset ___________________

temp_dir = '/tmp/tmp30m_3nsk'

    @pytest.fixture
    def dummy_data(temp_dir):
        # Create a dummy dataset
        dummy_data = [(np.random.rand(10, 3, 2), i % 3) for i in range(5)]
        data_path = os.path.join(temp_dir, 'data.npz')
>       np.savez(data_path, *dummy_data)

test/test_data_loader.py:25:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
../miniconda3/lib/python3.10/site-packages/numpy/lib/npyio.py:639: in savez
    _savez(file, args, kwds, False)
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

file = '/tmp/tmp30m_3nsk/data.npz'
args = ((array([[[0.69097819, 0.91616646],
        [0.33162476, 0.03605634],
        [0.56994853, 0.38930877]],

       [[0.4...775784]],

       [[0.96813509, 0.4836435 ],
        [0.75056261, 0.18810344],
        [0.93277804, 0.8280973 ]]]), 1))
kwds = {'arr_0': (array([[[0.69097819, 0.91616646],
        [0.33162476, 0.03605634],
        [0.56994853, 0.38930877]],

   ...7e-01, 1.44940277e-01],
        [8.74275614e-01, 9.13514123e-01],
        [9.54224020e-01, 6.93742218e-01]]]), 0), ...}
compress = False, allow_pickle = True, pickle_kwargs = None

    def _savez(file, args, kwds, compress, allow_pickle=True, pickle_kwargs=None):
        # Import is postponed to here since zipfile depends on gzip, an optional
        # component of the so-called standard library.
        import zipfile

        if not hasattr(file, 'write'):
            file = os_fspath(file)
            if not file.endswith('.npz'):
                file = file + '.npz'

        namedict = kwds
        for i, val in enumerate(args):
            key = 'arr_%d' % i
            if key in namedict.keys():
                raise ValueError(
                    "Cannot use un-named variables and keyword %s" % key)
            namedict[key] = val

        if compress:
            compression = zipfile.ZIP_DEFLATED
        else:
            compression = zipfile.ZIP_STORED

        zipf = zipfile_factory(file, mode="w", compression=compression)

        for key, val in namedict.items():
            fname = key + '.npy'
>           val = np.asanyarray(val)
E           ValueError: setting an array element with a sequence. The requested array has an inhomogeneous shape after 1 dimensions. The detected shape was (2,) + inhomogeneous part.

../miniconda3/lib/python3.10/site-packages/numpy/lib/npyio.py:740: ValueError
============================= short test summary info =============================
ERROR test/test_data_loader.py::test_samples_dataset - ValueError: setting an array element with a sequence. The requested array has ...
ERROR test/test_data_loader.py::test_trajectories_dataset - ValueError: setting an array element with a sequence. The requested array has ...
=========================== 4 passed, 2 errors in 2.36s ===========================
```
