### 2023-08-14 fc32a08
Add yes to installation with conda
