### 2023-08-14 5da8ed0
Use structured arrays to store positions and particle types
