### 2023-08-14 585a196
Use structured arrays to store positions and particle types (#53)
