### 2023-08-14 #5025 @kks32
Thanks @archermarx ! Ah...seems like the way `npz` is saved by numpy changed between Python versions of 3.9 and 3.10. I have now added the code to support both versions: https://github.com/geoelements/gns/pull/53

Could you please repull the gns repo and the test should now work. Thanks again for your patience.
