### 2023-08-14 geoelements/gns#54 @osorensen
Editorial comment on JOSS submission
I'm the editor of your JOSS submission, reading through the paper one final time. I will post suggested changes as issues in this repo.

First, running `@editorial check references` gives the following output:

```

OK DOIs

- 10.5281/zenodo.6658322 is OK
- 10.17603/ds2-0phb-dg64 is OK
- 10.1680/jgeot.15.LM.005 is OK

MISSING DOIs

- 10.1016/j.cma.2022.115704 may be a valid DOI for title: Model-Free Data-Driven Inference in Computational Mechanics
- 10.1145/3197517.3201293 may be a valid DOI for title: A Moving Least Squares Material Point Method with Displacement Discontinuity and Two-Way Rigid Body Coupling

INVALID DOIs

- 10.5555/3524938 is INVALID
```

Please correct the invalid DOI and add the two missing ones.

https://github.com/openjournals/joss-reviews/issues/5025
