### 2023-08-14 geoelements/gns#55 @osorensen
Editorial comment on JOSS submission
On line 29 there is a left parenthesis "(instead of ..." that does not seem to be closed anywhere with a right parenthesis:

https://github.com/openjournals/joss-reviews/issues/5025
