### 2023-08-14 geoelements/gns#56 @osorensen
Editorial comment on JOSS submission
Please remove the following line from `references.bib`

https://github.com/geoelements/gns/blob/585a196263b854027a8e2d2a2132fe3caaf52c76/references.bib#L145

The text "[Journal Article]" is added to the reference, and it should not be:

https://github.com/openjournals/joss-reviews/issues/5025
