### 2023-08-14 geoelements/gns#57 @osorensen
Editorial comment on JOSS submission
For consistency with the remaining references, use only a single set of curly braces also in this entry:

https://github.com/geoelements/gns/blob/585a196263b854027a8e2d2a2132fe3caaf52c76/references.bib#L46

https://github.com/openjournals/joss-reviews/issues/5025
