+++ references.bib
@@ -141,8 +141,7 @@ author={
    ISSN = {0016-8505
 1751-7656},
    DOI = {10.1680/jgeot.15.LM.005},
-   year = {2016},
-   type = {Journal Article}
+   year = {2016}
 }
 @article{hu2018mlsmpmcpic,
   title={A Moving Least Squares Material Point Method with Displacement Discontinuity and Two-Way Rigid Body Coupling},
