### 2023-08-14 ddc46c1
Fixes #57 paranthesis in references
