### 2023-08-14 6138d7f
Fixes #57 paranthesis in references
