+++ references.bib
@@ -43,9 +43,9 @@ year = {2022}
   publisher={IEEE}
 }
 @inproceedings{gilmer2017neural,
-  title={{Neural Message Passing for Quantum Chemistry}},
+  title={Neural Message Passing for Quantum Chemistry},
   author={Gilmer, Justin and Schoenholz, Samuel S and Riley, Patrick F and Vinyals, Oriol and Dahl, George E},
-  booktitle={{International Conference on Machine Learning}},
+  booktitle={International Conference on Machine Learning},
   pages={1263--1272},
   year={2017},
   organization={PMLR},
