+++ references.bib
@@ -28,7 +28,6 @@ year = {2022}
   booktitle={International Conference on Machine Learning},
   pages={8459--8468},
   year={2020},
-  doi={10.5555/3524938},
   url={https://dl.acm.org/doi/10.5555/3524938.3525722},
   organization={PMLR}
 }
@@ -67,6 +66,7 @@ year = {2022}
   title={Model-Free Data-Driven Inference in Computational Mechanics},
   author={Prume, Erik and Reese, Stefanie and Ortiz, Michael},
   journal={arXiv preprint arXiv:2207.06419},
+  doi={10.1016/j.cma.2022.115704},
   year={2022}
 }
 @software{vantassel2022gnsdata,
@@ -151,5 +151,6 @@ author={
   number={4},
   pages={150},
   year={2018},
+  doi={10.1145/3197517.3201293},
   publisher={ACM}
 }
