### 2023-08-15 441e8da
Update citation to v1.1.0
