### 2023-08-15 #5025 @kks32
GitHub GNS version [v1.1.0](https://github.com/geoelements/gns/releases/tag/v1.1.0)
Zenodo DOI: [10.5281/zenodo.8249671](https://doi.org/10.5281/zenodo.8249671)
