### 2023-08-15 #5025 @osorensen
@kks32, regarding this point:

>- [ ] Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

Could you please change the title of the archive so it exactly matches the JOSS paper. That is, change it to "GNS: A generalizable Graph Neural Network-based simulator for particulate and fluid modeling".
