+++ CITATION.cff
@@ -9,6 +9,6 @@ authors:
     orcid: https://orcid.org/0000-0002-1601-3354      
 title: "GNS: A generalizable Graph Neural Network-based simulator for particulate and fluid modeling"
 version: v1.1.1
-doi: 10.5281/zenodo.8249671
+doi: 10.5281/zenodo.8249813 
 date-released: 2023-08-15
 url: "https://github.com/geoelements/gns"
