### 2023-08-16 58dc74f
add a doc for explaining details about training data
