### 2023-08-18 75026c5
add a doc for explaining details about training data (#59)
* add a doc for explaining details about training data

* Add image

* Change heading levels
