### 2023-08-18 4d83d69
Add MeshNet to _sidebar.md
