+++ docs/_sidebar.md
@@ -1,2 +1,3 @@
 <!-- docs/_sidebar.md -->
 * [GNS Theory](theory.md)
+* [MeshNet data](flow_data.md)
