### 2023-08-21 8e2f451
bug fix for cpu and gpu conditioning
