### 2023-08-21 d7b9290
include loss in training_state file
