### 2023-08-22 ad4b7c2
Use rank if cuda, else use device (which is cpu), to deal with different behaviors depending on cpu and gpu.
